requirements.txt
Django==1.7
Fabric==1.10.0
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.6.1
PyYAML==3.11
WebOb==1.4
ansible==1.7.2
appenlight-client==0.6.10
django-appconf==0.6
django-compressor==1.4
django-debug-toolbar==1.2.1
django-ipware==0.1.0
django-localflavor==1.0
django-mailgun==0.2.2
ecdsa==0.11
newrelic==2.34.0.29
paramiko==1.15.1
psycopg2==2.5.4
pycrypto==2.6.1
raven==5.1.1
requests==2.4.3
six==1.8.0
sqlparse==0.1.13
wsgiref==0.1.2
